examples: add path validation

+++ examples/writable/writable.go
@@ -19,19 +19,38 @@ func NewWritableDirFS(path string) *writableDirFS {
 }
 
 func (fsys *writableDirFS) OpenWriter(name string, flag int) (io.WriteCloser, error) {
+	if !fs.ValidPath(name) {
+		return nil, &fs.PathError{Op: "open", Path: name, Err: fs.ErrInvalid}
+	}
 	return os.OpenFile(path.Join(fsys.path, name), flag, fs.ModePerm)
 }
 
 func (fsys *writableDirFS) Truncate(name string, size int64) error {
+	if !fs.ValidPath(name) {
+		return &fs.PathError{Op: "truncate", Path: name, Err: fs.ErrInvalid}
+	}
 	return os.Truncate(path.Join(fsys.path, name), size)
 }
 
 func (fsys *writableDirFS) Remove(name string) error {
+	if !fs.ValidPath(name) {
+		return &fs.PathError{Op: "remove", Path: name, Err: fs.ErrInvalid}
+	}
 	return os.Remove(path.Join(fsys.path, name))
 }
 
 func (fsys *writableDirFS) Mkdir(name string, mode fs.FileMode) error {
+	if !fs.ValidPath(name) {
+		return &fs.PathError{Op: "mkdir", Path: name, Err: fs.ErrInvalid}
+	}
 	return os.Mkdir(path.Join(fsys.path, name), mode)
+}
+
+func (fsys *writableDirFS) Rename(name, newName string) error {
+	if !fs.ValidPath(name) || !fs.ValidPath(newName) {
+		return &fs.PathError{Op: "rename", Path: name, Err: fs.ErrInvalid}
+	}
+	return os.Rename(path.Join(fsys.path, name), path.Join(fsys.path, newName))
 }
 
 func main() {
@@ -45,7 +64,10 @@ func main() {
 		mountPoint = os.Args[2]
 	}
 
-	mount, _ := fsmount.MountFS(mountPoint, NewWritableDirFS(srcDir), nil)
+	mount, err := fsmount.MountFS(mountPoint, NewWritableDirFS(srcDir), nil)
+	if err != nil {
+		panic(err)
+	}
 	defer mount.Close()
 
 	// Block forever
